requirements.txt
# Основные библиотеки
langchain==0.1.5
langchain-community==0.0.16
langchain-huggingface==0.0.8
langchain-openai==0.0.5
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3

# Векторная база данных
qdrant-client==1.7.0

# Модели и эмбединги
transformers==4.37.2
sentence-transformers==2.2.2
fastembed==0.1.1
bitsandbytes==0.42.0
accelerate==0.26.1
torch==2.1.2
safetensors==0.4.1

# Утилиты
tqdm==4.66.1
python-dotenv==1.0.0